requirements.txt
pyyaml>=6.0.2
pytest>=8.2.0
click>=8.1.7